refactor(runtime): extract resource page context loading
- move GenerateDesignatorContext and designator context loading into runtime/weave/page_contexts.ts
- move child identifier, RDF type hint, Knop owner title, artifact link, and context history helpers with that context boundary
- keep weave.ts focused on generate/page assembly orchestration
- update codebase overview for the new runtime page-context module
- verify with check, lint, focused integration tests, e2e weave CLI tests, and import graph cycle scan
