+++ documentation/notes/wd.codebase-overview.md
@@ -27,7 +27,7 @@ created: 1773673181726
   `runtime/config` now carries the first default effective-config seam: it loads Weave default RDF, resolves artifact-role history and ResourcePage policies, parses historical ResourcePage regeneration policy, parses default payload history/state/manifestation naming policies, parses the default config-resolution profile, and models first-pass Knop inherited-config propagation controls without changing fixture-backed behavior yet
   runtime inventory discovery, workspace loaders, and page rendering now carry the root designator path as a first-class resource when a root Knop exists at `_knop`
   runtime weave orchestration prepares a `PreparedWeaveExecution` once from shared target semantics, then dispatches version planning/writes and page generation from that prepared object instead of re-deriving generate targets from the raw request
-  `runtime/weave/weave.ts` remains the public façade for validate/version/generate/weave operations, while `runtime/weave/errors.ts`, `runtime/weave/mesh_state.ts`, `runtime/weave/planning_context.ts`, `runtime/weave/artifact_loaders.ts`, `runtime/weave/candidate_loader.ts`, `runtime/weave/request_normalization.ts`, `runtime/weave/prepared_execution.ts`, `runtime/weave/version_execution.ts`, `runtime/weave/execution_config.ts`, `runtime/weave/workspace_paths.ts`, `runtime/weave/progress.ts`, `runtime/weave/raw_source_panels.ts`, and `runtime/weave/timing_helpers.ts` now hold shared runtime errors, mesh-state loading, command-scoped staged filesystem/cache behavior, artifact-loading helpers used by both candidate loading and page context loading, candidate discovery/coverage checks, request normalization, prepared execution, version planning/writes, effective config policy adaptation, workspace path conversion, progress types, ResourcePage raw-source panel loading, and optional timing wrappers. The planning context owns the overlay-aware read cache, dependency-aware candidate cache, planned-file staging, and cache counters used by `WEAVE_TIMING`.
+  `runtime/weave/weave.ts` remains the public façade for validate/version/generate/weave operations, while `runtime/weave/errors.ts`, `runtime/weave/mesh_state.ts`, `runtime/weave/planning_context.ts`, `runtime/weave/artifact_loaders.ts`, `runtime/weave/candidate_loader.ts`, `runtime/weave/request_normalization.ts`, `runtime/weave/prepared_execution.ts`, `runtime/weave/version_execution.ts`, `runtime/weave/execution_config.ts`, `runtime/weave/workspace_paths.ts`, `runtime/weave/progress.ts`, `runtime/weave/page_contexts.ts`, `runtime/weave/raw_source_panels.ts`, and `runtime/weave/timing_helpers.ts` now hold shared runtime errors, mesh-state loading, command-scoped staged filesystem/cache behavior, artifact-loading helpers used by both candidate loading and page context loading, candidate discovery/coverage checks, request normalization, prepared execution, version planning/writes, effective config policy adaptation, workspace path conversion, progress types, ResourcePage designator-context loading, ResourcePage raw-source panel loading, and optional timing wrappers. The planning context owns the overlay-aware read cache, dependency-aware candidate cache, planned-file staging, and cache counters used by `WEAVE_TIMING`.
   runtime weave planning now passes Weave's default effective support-history and payload naming policies into version planning, letting first Knop and first payload weave outputs keep `_knop/_meta` current-only while preserving payload and inventory history behavior and keeping ordinal payload paths as the configured default
   runtime weave planning now passes current MeshMetadata into version planning so core can use `_mesh/_meta` as the MeshInventory progression source for the first `_mesh/_meta` migration seam
   runtime page generation and versioned inventory rendering now filter `sflo:hasResourcePage` candidates through the effective resource-page generation policy by owning artifact role, using stable history/state membership rather than mutable current/latest pointers, so `generate`, `suppress`, `defer`, and explicit-target `onRequest` have a materialization seam independent of history policy without leaving suppressed page promises in inventory
